Move all allocators to dedicated files in new allocators/ folder
