+++ src/clean-core/allocator.hh
@@ -70,246 +70,9 @@ struct allocator
     virtual ~allocator() = default;
 };
 
-
 /// global instance of the system allocator (malloc / free) (thread safe)
+/// see allocators/system_allocator.cc for implementation
 extern allocator* const system_allocator;
-
-/// trivial linear allocator operating in a given buffer
-/// cannot free individual allocations, only reset entirely
-struct linear_allocator final : allocator
-{
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override
-    {
-        CC_ASSERT(_buffer_begin != nullptr && "linear_allocator uninitialized");
-
-        auto* const padded_res = cc::align_up(_head, align);
-        CC_ASSERT(padded_res + size <= _buffer_end && "linear_allocator overcommitted");
-        _head = padded_res + size;
-
-        _latest_allocation = padded_res;
-        return padded_res;
-    }
-
-    void free(void* ptr) override
-    {
-        // no-op
-        (void)ptr;
-    }
-
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override
-    {
-        if (!ptr || ptr != _latest_allocation)
-        {
-            return new_size > 0 ? this->alloc(new_size, align) : nullptr;
-        }
-
-        std::byte* const ptr_byte = static_cast<std::byte*>(ptr);
-        CC_ASSERT(old_size == _head - ptr_byte && "old_size incorrect");
-        CC_ASSERT(ptr_byte + new_size <= _buffer_end && "linear_allocator overcommitted");
-
-        _head = ptr_byte + new_size;
-        return ptr_byte;
-    }
-
-    void reset()
-    {
-        _head = _buffer_begin;
-        _latest_allocation = nullptr;
-    }
-
-    size_t allocated_size() const { return _head - _buffer_begin; }
-    size_t remaining_size() const { return _buffer_end - _head; }
-    size_t max_size() const { return _buffer_end - _buffer_begin; }
-    float allocated_ratio() const { return allocated_size() / float(max_size()); }
-
-    std::byte* buffer() const { return _buffer_begin; }
-
-    linear_allocator() = default;
-    explicit linear_allocator(span<std::byte> buffer) : _buffer_begin(buffer.data()), _head(buffer.data()), _buffer_end(buffer.data() + buffer.size())
-    {
-    }
-
-private:
-    std::byte* _buffer_begin = nullptr;
-    std::byte* _head = nullptr;
-    std::byte* _buffer_end = nullptr;
-    std::byte* _latest_allocation = nullptr;
-};
-
-
-/// stack allocator operating in a given buffer
-/// like a linear allocator, but can also free the most recent allocation
-struct stack_allocator final : allocator
-{
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override;
-
-    /// NOTE: ptr must be the most recent allocation received
-    void free(void* ptr) override;
-
-    /// NOTE: ptr must be the most recent allocation received
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override;
-
-    void reset()
-    {
-        _head = _buffer_begin;
-        _last_alloc_id = 0;
-    }
-
-    stack_allocator() = default;
-    explicit stack_allocator(span<std::byte> buffer) : _buffer_begin(buffer.data()), _head(buffer.data()), _buffer_end(buffer.data() + buffer.size())
-    {
-    }
-
-private:
-    std::byte* _buffer_begin = nullptr;
-    std::byte* _head = nullptr;
-    std::byte* _buffer_end = nullptr;
-    int32_t _last_alloc_id = 0;
-};
-
-/// linear allocator operating in virtual memory
-/// reserves pages on init, commits pages on demand
-/// only frees pages if explicitly called
-struct virtual_linear_allocator final : allocator
-{
-    virtual_linear_allocator() = default;
-    explicit virtual_linear_allocator(size_t max_size_bytes, size_t chunk_size_bytes = 65536) { initialize(max_size_bytes, chunk_size_bytes); }
-    ~virtual_linear_allocator() { destroy(); }
-
-    // max_size_bytes: amount of contiguous virtual memory being reserved
-    // chunk_size_bytes: increment of physical memory being committed whenever more is required
-    // note there is a lower limit on virtual allocation granularity (Win32: 64K = 16 pages)
-    void initialize(size_t max_size_bytes, size_t chunk_size_bytes = 65536);
-
-    void destroy();
-
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override;
-
-    void free(void* ptr) override { (void)ptr; }
-
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override;
-
-    // free all current allocations
-    // does not decommit any memory!
-    size_t reset()
-    {
-        size_t const num_bytes_allocated = _physical_current - _virtual_begin;
-        _physical_current = _virtual_begin;
-        _last_allocation = nullptr;
-        return num_bytes_allocated;
-    }
-
-    // decommit the physical memory of all pages not currently required
-    // returns amount of bytes decommitted
-    size_t decommit_idle_memory();
-
-    // amount of bytes in the virtual address range
-    size_t get_virtual_size_bytes() const { return _virtual_end - _virtual_begin; }
-
-    // amount of bytes in the physically committed memory
-    size_t get_physical_size_bytes() const { return _physical_end - _virtual_begin; }
-
-    // amount of bytes in the physically committed and allocated memory
-    size_t get_allocated_size_bytes() const { return _physical_current - _virtual_begin; }
-
-private:
-    std::byte* _virtual_begin = nullptr;
-    std::byte* _virtual_end = nullptr;
-    std::byte* _physical_current = nullptr;
-    std::byte* _physical_end = nullptr;
-    std::byte* _last_allocation = nullptr;
-    size_t _chunk_size_bytes = 0;
-};
-
-/// stack allocator operating in virtual memory
-/// reserves pages on init, commits pages on demand
-/// only frees pages if explicitly called
-struct virtual_stack_allocator final : allocator
-{
-    virtual_stack_allocator() = default;
-    explicit virtual_stack_allocator(size_t max_size_bytes, size_t chunk_size_bytes = 65536) { initialize(max_size_bytes, chunk_size_bytes); }
-    ~virtual_stack_allocator() { destroy(); }
-
-    // max_size_bytes: amount of contiguous virtual memory being reserved
-    // chunk_size_bytes: increment of physical memory being committed whenever more is required
-    // note there is a lower limit on virtual allocation granularity (Win32: 64K = 16 pages)
-    void initialize(size_t max_size_bytes, size_t chunk_size_bytes = 65536);
-
-    void destroy();
-
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override;
-
-    /// NOTE: ptr must be the most recent allocation received
-    void free(void* ptr) override;
-
-    /// NOTE: ptr must be the most recent allocation received
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override;
-
-    // free all current allocations
-    // does not decommit any memory!
-    size_t reset();
-
-    // decommit the physical memory of all pages not currently required
-    // returns amount of bytes decommitted
-    size_t decommit_idle_memory();
-
-    // amount of bytes in the virtual address range
-    size_t get_virtual_size_bytes() const { return _virtual_end - _virtual_begin; }
-
-    // amount of bytes in the physically committed memory
-    size_t get_physical_size_bytes() const { return _physical_end - _virtual_begin; }
-
-    // amount of bytes in the physically committed and allocated memory
-    size_t get_allocated_size_bytes() const { return _physical_current - _virtual_begin; }
-
-    // returns whether the given ptr is the latest allocation, meaning it can be freed or reallocated
-    bool is_latest_allocation(void* ptr) const;
-
-private:
-    std::byte* _virtual_begin = nullptr;
-    std::byte* _virtual_end = nullptr;
-    std::byte* _physical_current = nullptr;
-    std::byte* _physical_end = nullptr;
-    int32_t _last_alloc_id = 0;
-    size_t _chunk_size_bytes = 0;
-};
-
-/// Two Level Segregated Fit allocator
-/// O(1) cost for alloc, free, realloc
-/// extremely low memory overhead, 4 byte per allocation
-struct tlsf_allocator final : allocator
-{
-    tlsf_allocator() = default;
-    explicit tlsf_allocator(cc::span<std::byte> buffer) { initialize(buffer); }
-    ~tlsf_allocator() { destroy(); }
-
-    void initialize(cc::span<std::byte> buffer);
-    void destroy();
-
-    // provide an additional memory pool to the TLSF (can be done multiple times)
-    void add_pool(cc::span<std::byte> buffer);
-
-    // returns false if internal consistency checks fail
-    bool check_consistency();
-
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override;
-
-    void free(void* ptr) override;
-
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override;
-
-    tlsf_allocator(tlsf_allocator&& rhs) noexcept : _tlsf(rhs._tlsf) { rhs._tlsf = nullptr; }
-    tlsf_allocator& operator==(tlsf_allocator&& rhs) noexcept
-    {
-        destroy();
-        _tlsf = rhs._tlsf;
-        rhs._tlsf = nullptr;
-        return *this;
-    }
-
-private:
-    void* _tlsf = nullptr;
-};
 
 //
 // templated virtual allocator interface implementation below
