+++ src/clean-core/allocators/atomic_linear_allocator.cc
@@ -1,0 +1,1 @@
+#include "atomic_linear_allocator.hh"
