+++ src/clean-core/allocators/linear_allocator.cc
@@ -1,0 +1,1 @@
+#include "linear_allocator.hh"
