+++ src/clean-core/allocators/synced_tlsf_allocator.cc
@@ -1,0 +1,1 @@
+#include "synced_tlsf_allocator.hh"
