+++ src/clean-core/allocators/synced_virtual_linear_allocator.cc
@@ -1,0 +1,1 @@
+#include "synced_virtual_linear_allocator.hh"
