--- src/clean-core/threadsafe_allocators.hh
@@ -1,243 +1,0 @@
-#pragma once
-
-#include <cstddef>
-
-#include <atomic>
-
-#include <clean-core/allocator.hh>
-#include <clean-core/new.hh>
-#include <clean-core/utility.hh>
-#include <clean-core/spin_lock.hh>
-
-namespace cc
-{
-/// thread safe pool allocator, O(1) alloc and free
-/// cannot alloc buffers > block_size (ie. does not search contiguous blocks)
-/// provided buffer and block size must be aligned to a multiple of all requests (this is verified)
-struct atomic_pool_allocator final : allocator
-{
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override
-    {
-        if (size > _block_size)
-            return nullptr;
-
-        CC_ASSERT(!is_full() && "pool_allocator full");
-
-        // CAS-loop to acquire a free node and write the matching next pointer
-        bool cas_success = false;
-        std::byte* acquired_node = nullptr;
-        do
-        {
-            // acquire-candidate: the current value of _first_free_node
-            acquired_node = _first_free_node.load(std::memory_order_acquire);
-            // read the in-place next pointer of this node
-            std::byte* const next_pointer_of_acquired = *reinterpret_cast<std::byte**>(acquired_node);
-
-            // compare-exchange these two - spurious failure if raced
-            cas_success = std::atomic_compare_exchange_weak_explicit(&_first_free_node, &acquired_node, next_pointer_of_acquired,
-                                                                     std::memory_order_seq_cst, std::memory_order_relaxed);
-        } while (!cas_success);
-
-        CC_ASSERT(cc::is_aligned(acquired_node, align) && "pool buffer and blocks must be aligned to a multiple of all requests");
-        return acquired_node;
-    }
-
-    void free(void* ptr) override
-    {
-        if (!ptr)
-            return;
-
-        std::byte* const freed_node = static_cast<std::byte*>(ptr);
-        CC_ASSERT(freed_node >= _buffer_begin && freed_node - _buffer_begin <= ptrdiff_t(_buffer_size) && "pointer in pool_allocator::free is not part of the buffer");
-        CC_ASSERT((freed_node - _buffer_begin) % _block_size == 0 && "freed pointer is not on a node boundary");
-
-        // write the in-place next pointer of this node
-        bool cas_success = false;
-        do
-        {
-            std::byte* expected_first_free = _first_free_node.load(std::memory_order_acquire);
-
-            // write the in-place next pointer of this node provisionally
-            new (cc::placement_new, freed_node) std::byte*(expected_first_free);
-
-            // CAS write the newly released node if the expected wasn't raced
-            cas_success = std::atomic_compare_exchange_weak_explicit(&_first_free_node, &expected_first_free, freed_node, std::memory_order_seq_cst,
-                                                                     std::memory_order_relaxed);
-        } while (!cas_success);
-    }
-
-    /// returns the offset of a node to the buffer start in bytes
-    size_t get_node_offset_bytes(void const* ptr) const
-    {
-        std::byte const* const node = static_cast<std::byte const*>(ptr);
-        CC_ASSERT(node >= _buffer_begin && node - _buffer_begin <= ptrdiff_t(_buffer_size)
-                  && "pointer in pool_allocator::get_node_offset_bytes is not part of the buffer");
-        CC_ASSERT((node - _buffer_begin) % _block_size == 0 && "pointer is not on a node boundary");
-        return size_t(node - _buffer_begin);
-    }
-
-    size_t get_node_index(void const* ptr) const
-    {
-        auto const offset_bytes = get_node_offset_bytes(ptr);
-        return offset_bytes / _block_size;
-    }
-
-    bool is_full() const { return _first_free_node == nullptr; }
-    size_t max_size_bytes() const { return _buffer_size; }
-    size_t block_size_bytes() const { return _block_size; }
-    size_t max_num_blocks() const { return _buffer_size / _block_size; }
-
-    atomic_pool_allocator() = default;
-    explicit atomic_pool_allocator(span<std::byte> buffer, size_t block_size);
-
-    void initialize(span<std::byte> buffer, size_t block_size);
-
-    template <class T>
-    void initialize_from_array(span<T> elements, size_t block_size_in_elements = 1)
-    {
-        return initialize(cc::as_byte_span(elements), sizeof(T) * block_size_in_elements);
-    }
-
-private:
-    std::byte* _buffer_begin = nullptr;
-    std::atomic<std::byte*> _first_free_node = nullptr;
-    size_t _buffer_size = 0;
-    size_t _block_size = 0;
-};
-
-/// thread safe version of cc::linear_allocator
-struct atomic_linear_allocator final : allocator
-{
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override
-    {
-        CC_ASSERT(_buffer_begin != nullptr && "atomic_linear_allocator unintialized");
-
-        auto const buffer_size = size + align - 1; // align worst case buffer to satisfy up-aligning
-        auto const buffer_start = _offset.fetch_add(buffer_size, std::memory_order_acquire);
-
-        auto* const alloc_start = _buffer_begin + buffer_start;
-        auto* const padded_res = cc::align_up(alloc_start, align);
-
-        CC_ASSERT(padded_res - alloc_start < std::ptrdiff_t(align) && "up-align OOB");
-        CC_ASSERT(padded_res + size <= _buffer_end && "atomic_linear_allocator overcommitted");
-
-        return padded_res;
-    }
-
-    void free(void* ptr) override
-    {
-        // no-op
-        (void)ptr;
-    }
-
-    void reset() { _offset.store(0, std::memory_order_release); }
-
-    size_t allocated_size() const { return _offset.load(); }
-    size_t max_size() const { return _buffer_end - _buffer_begin; }
-    float allocated_ratio() const { return allocated_size() / float(max_size()); }
-
-    atomic_linear_allocator() = default;
-    explicit atomic_linear_allocator(span<std::byte> buffer) : _buffer_begin(buffer.data()), _offset(0), _buffer_end(buffer.data() + buffer.size()) {}
-
-    void initialize(span<std::byte> buffer)
-    {
-        // atomics cant be moved, making this necessary
-        _buffer_begin = buffer.data();
-        _offset = 0;
-        _buffer_end = buffer.data() + buffer.size();
-    }
-
-private:
-    std::byte* _buffer_begin = nullptr;
-    std::atomic<std::size_t> _offset = {0};
-    std::byte* _buffer_end = nullptr;
-};
-
-/// Synchronized (mutexed) version of tlsf_allocator
-template <class MutexT = cc::spin_lock>
-struct synced_tlsf_allocator final : allocator
-{
-    synced_tlsf_allocator() = default;
-    explicit synced_tlsf_allocator(cc::span<std::byte> buffer) : _backing(buffer) {}
-    ~synced_tlsf_allocator() { _backing.destroy(); }
-
-
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.alloc(size, align);
-    }
-
-    void free(void* ptr) override
-    {
-        auto lg = cc::lock_guard(_mutex);
-        _backing.free(ptr);
-    }
-
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.realloc(ptr, old_size, new_size, align);
-    }
-
-    void initialize(cc::span<std::byte> buffer) { _backing.initialize(buffer); }
-    void destroy() { _backing.destroy(); }
-
-private:
-    MutexT _mutex;
-    cc::tlsf_allocator _backing;
-};
-
-/// Synchronized (mutexed) version of virtual_linear_allocator
-template <class MutexT = cc::spin_lock>
-struct synced_virtual_linear_allocator final : allocator
-{
-    synced_virtual_linear_allocator() = default;
-    explicit synced_virtual_linear_allocator(size_t max_size_bytes, size_t chunk_size_bytes = 65536)
-    {
-        _backing.initialize(max_size_bytes, chunk_size_bytes);
-    }
-
-    void initialize(size_t max_size_bytes, size_t chunk_size_bytes = 65536) { _backing.initialize(max_size_bytes, chunk_size_bytes); }
-
-    void destroy()
-    {
-        auto lg = cc::lock_guard(_mutex);
-        _backing.destroy();
-    }
-
-    std::byte* alloc(size_t size, size_t align = alignof(std::max_align_t)) override
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.alloc(size, align);
-    }
-
-    void free(void*) override {} // nothing
-
-    std::byte* realloc(void* ptr, size_t old_size, size_t new_size, size_t align = alignof(std::max_align_t)) override
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.realloc(ptr, old_size, new_size, align);
-    }
-
-    size_t reset()
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.reset();
-    }
-
-    // decommit the physical memory of all pages not currently required
-    // returns amount of bytes decommitted
-    size_t decommit_idle_memory()
-    {
-        auto lg = cc::lock_guard(_mutex);
-        return _backing.decommit_idle_memory();
-    }
-
-    cc::virtual_linear_allocator const& get_backing() const { return _backing; }
-
-private:
-    MutexT _mutex;
-    cc::virtual_linear_allocator _backing;
-};
-}
